scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0
openai>=1.3.0
pydantic>=2.5.0
click>=8.1.0
colorama>=0.4.6
python-dotenv>=1.0.0
pytest>=7.4.0

# Optional: For WebDriver/Selenium engine mode
selenium>=4.15.0
webdriver-manager>=4.0.0
